% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropped_df.R
\name{dropped_df}
\alias{dropped_df}
\title{Display the treatment cases that will be dropped under a particular matched sample}
\usage{
dropped_df(matched_df, use_attr = TRUE, cov_cols = NULL)
}
\arguments{
\item{matched_df}{A data frame output by \code{\link[=matched_sample]{matched_sample()}}. This should contain columns for \code{arm_col}, all elements of \code{cov_cols} and \code{include}. It must have the \code{Arm} column with levels \code{"Intervention"} and \code{"Comparison"}. It may also contain other columns.}

\item{use_attr}{logical. Should the attributes of \code{matched_df} (set by \code{\link[=matched_sample]{matched_sample()}}) be used to specify \code{cov_cols}?}

\item{cov_cols}{(only specify if \code{use_attr == FALSE}) A character vector containing the column names of the covariates to be matched on. These should all be factor / categorical data.}
}
\value{
\code{\link[=dropped_df]{dropped_df()}} returns a data frame with the following columns
\itemize{
\item A column for each of \code{cov_cols}
\item \code{n_dropped}: how many treated / intervention cases with this combination of \code{cov_cols} have been dropped / downsampled
\item \code{n_kept}: how many treated / intervention cases with this combination of \code{cov_cols} have been kept
The rows will be ordered in descending order of \code{n_dropped}
}
}
\description{
Display the treatment cases that will be dropped under a particular matched sample
}
\examples{
## First use `propscore_df` to create a propensity score data frame
## This example uses the joint approach and specifies p(Z=1) such
## that the expected size of the comparison group is 300

eg_j_exp300 = propscore_df(
   df = eg_data,
   cov_cols = c("risk_ass", "category", "sus_age_bin"),
   arm_col = "Arm",
   intervention_level = "Intervention",
   cov_dist = "joint",
   pz1_fn = pz1_expectedN(300)
   )

## Now we can use `eg_j_exp300` to find a matched comparison group
## In this version we sample from the comparison group with replacement
## and do not drop any intervention cases for which there are no
## comparison cases to sample

match_j_exp300_replace = matched_sample(
   df = eg_data,
   propscore_df = eg_j_exp300,
   seed = 20,
   replace = TRUE,
   downsample = FALSE,
   drop_int = FALSE
   )

## Finally we use [dropped_df()] to summarise the treated / intervention
## cases that have been dropped or downsampled
## included `include` times

match_j_exp300_replace |> dropped_df()

}
\seealso{
\code{\link[=matched_sample]{matched_sample()}}
}
