% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss_functions.R
\name{pss.rc}
\alias{pss.rc}
\title{PSS for rows and columns}
\usage{
pss.rc(samp, phyl.rows, phyl.cols, q.rows, q.cols)
}
\arguments{
\item{samp}{Required. A matrix. Elements of the matrix are interaction frequencies.}

\item{phyl.rows}{Required. An object of class phylo. This is a phylogenetic tree that includes all the taxa in the rows of samp. This tree may have taxa not included in samp.}

\item{phyl.cols}{Required. An object of class phylo. This is a phylogenetic tree that includes all the taxa in the columns of samp. This tree may have taxa not included in samp.}

\item{q.rows}{Optional. A named numerical vector with the relative availabilities (i.e., sum(q) = 1) of the taxa in the rows of samp. Default is to estimate these values from interaction frequencies.}

\item{q.cols}{Optional. A named numerical vector with the relative availabilities (i.e., sum(q) = 1) of the taxa in the columns of samp. Default is to estimate these values from interaction frequencies.}
}
\value{
A list with two dataframes (rows and columns) with results for each species in samp.
}
\description{
PSS for the rows and columns of an interaction matrix. PSS is a standardized effect size (SES) of klMPD.
}
\author{
Carlos J. Pardo De la Hoz
}
