% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss_functions.R
\name{ses.mpd.rc}
\alias{ses.mpd.rc}
\title{Standardized Effect Size of wMPD for rows and columns}
\usage{
ses.mpd.rc(samp, phyl.rows, phyl.cols)
}
\arguments{
\item{samp}{Required. A matrix. Elements of the matrix are interaction frequencies.}

\item{phyl.rows}{Required. An object of class phylo. This is a phylogenetic tree that includes all the taxa in the rows of samp. This tree may have taxa not included in samp.}

\item{phyl.cols}{Required. An object of class phylo. This is a phylogenetic tree that includes all the taxa in the columns of samp. This tree may have taxa not included in samp.}
}
\value{
A list with two dataframes (rows and columns) with results for each taxon in samp. See ?ses.mpd in Picante for more details.
}
\description{
Standardized Effect Size of the interaction-frequency-weighted mean pairwise phylogenetic distance (wMPD) for the taxa in the rows and columns of an interaction matrix. This is simply an extension of picante's function ses.mpd
}
\author{
Carlos J. Pardo De la Hoz
}
