% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Create a summary table for a vector of seeds}
\usage{
summary_table(
  df,
  propscore_df,
  seeds,
  smd_cols,
  smd_order,
  replace = TRUE,
  downsample = FALSE,
  drop_int = FALSE,
  use_attr = TRUE,
  cov_cols = NULL,
  arm_col = NULL,
  intervention_level = NULL
)
}
\arguments{
\item{df}{A data frame from which to generate matched samples}

\item{propscore_df}{A data frame (created by \code{\link[=propscore_df]{propscore_df()}}, or with the same format) giving a propensity score (in column \code{PropScore}) for every covariate being matched on}

\item{seeds}{A vector of random seeds to be used}

\item{smd_cols}{A character vector of columns for which to assess balance. This can be the same as \code{cov_cols} but it doens't have to be.}

\item{smd_order}{An integer denoting the highest order of interaction for which to calculate the standardised mean difference (SMD)}

\item{replace}{logical. Passed to \code{\link[=matched_sample]{matched_sample()}}. Should the comparison cases be sampled with replacement?}

\item{downsample}{logical. Passed to \code{\link[=matched_sample]{matched_sample()}}. Should the treatment cases be downsampled when there aren't enough comparison cases? Only one of \code{replace} and \code{downsample} should be TRUE.}

\item{drop_int}{logical. Passed to \code{\link[=matched_sample]{matched_sample()}}. Should cases in the intervention/treated group be dropped if there are no equivalent cases in the comparison cohort to sample from?}

\item{use_attr}{logical. Should the attributes of \code{propscore_df} (set by \code{\link[=propscore_df]{propscore_df()}}) be used to specify \code{cov_cols}, \code{arm_col} and \code{intervention_level}?}

\item{cov_cols}{(only specify if \code{use_attr == FALSE}) A character vector containing the column names of the covariates to be matched on. These should all be factor / categorical data.}

\item{arm_col}{(only specify if \code{use_attr == FALSE}) The name of the column indicating which rows are treated cases and which are comparison cases. These should be factor or character, with only two levels / options.}

\item{intervention_level}{(only specify if \code{use_attr == FALSE}) The value in the \code{arm_col} for the treatment cases}
}
\value{
\code{\link[=summary_table]{summary_table()}} returns a data frame with a row for each random seed and columns
\itemize{
\item seed
\item \code{n_int}: the number of treated / intervention cases included in the matched sample for that seed
\item \code{n_comp}: the number of comparison cases included in the matched sample for that seed
\item \code{SMD_...}: column for each covariate and covariate interaction, giving the calculated SMD
}
}
\description{
Create a summary table for a vector of seeds
}
\examples{
## First use `propscore_df` to create a propensity score data frame
## This example uses the joint approach and specifies p(Z=1) such
## that the expected size of the comparison group is 300

eg_j_exp300 = propscore_df(
   df = eg_data,
   cov_cols = c("risk_ass", "category", "sus_age_bin"),
   arm_col = "Arm",
   intervention_level = "Intervention",
   cov_dist = "joint",
   pz1_fn = pz1_expectedN(300)
   )

## Now we can use `eg_j_exp300` to find a matched comparison group
## In this version we sample from the comparison group with replacement
## and do not drop any intervention cases for which there are no
## comparison cases to sample

match_j_exp300_replace = matched_sample(
   df = eg_data,
   propscore_df = eg_j_exp300,
   seed = 20,
   replace = TRUE,
   downsample = FALSE,
   drop_int = FALSE
   )

## Finally we use `expand_matched_df` to expand this so that each row is
## included `include` times

match_j_exp300_replace |> expand_matched_df()

}
\seealso{
\code{\link[=matched_sample]{matched_sample()}}, \code{\link[=expand_matched_df]{expand_matched_df()}}
}
