% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_matched_df.R
\name{expand_matched_df}
\alias{expand_matched_df}
\title{Expand a matched comparison data frame}
\usage{
expand_matched_df(matched_df)
}
\arguments{
\item{matched_df}{A data frame output by \code{\link[=matched_sample]{matched_sample()}}. This should contain columns for \code{arm_col}, all elements of \code{cov_cols} and \code{include}. It can also contain other columns.}
}
\value{
\code{\link[=expand_matched_df]{expand_matched_df()}} returns a data frame with all the columns of \code{matched_df} except include. Each row (row \code{i}) is now included \code{include[i]} times. That is, rows with \code{matched_df$include[i]==0} are omitted from the returned data frame, and rows with \code{matched_df$include[i] > 1} are included multiple times.
}
\description{
Expand a matched comparison data frame
}
\examples{
## First use `propscore_df` to create a propensity score data frame
## This example uses the joint approach and specifies p(Z=1) such
## that the expected size of the comparison group is 300

eg_j_exp300 = propscore_df(
   df = eg_data,
   cov_cols = c("risk_ass", "category", "sus_age_bin"),
   arm_col = "Arm",
   intervention_level = "Intervention",
   cov_dist = "joint",
   pz1_fn = pz1_expectedN(300)
   )

## Now we can use `eg_j_exp300` to find a matched comparison group
## In this version we sample from the comparison group with replacement
## and do not drop any intervention cases for which there are no
## comparison cases to sample

match_j_exp300_replace = matched_sample(
   df = eg_data,
   propscore_df = eg_j_exp300,
   seed = 20,
   replace = TRUE,
   downsample = FALSE,
   drop_int = FALSE
   )

## Finally we use `expand_matched_df` to expand this so that each row is
## included `include` times

match_j_exp300_replace |> expand_matched_df()

}
\seealso{
\code{\link[=matched_sample]{matched_sample()}}
}
