% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_sample.R
\name{matched_sample}
\alias{matched_sample}
\title{Form a matched sample}
\usage{
matched_sample(
  df,
  propscore_df,
  seed,
  use_attr = TRUE,
  replace = TRUE,
  downsample = FALSE,
  drop_int = FALSE,
  cov_cols = NULL,
  arm_col = NULL,
  intervention_level = NULL
)
}
\arguments{
\item{df}{A data frame containing treatment group and comparison cohort data}

\item{propscore_df}{A data frame with a column for each variable in \code{cov_cols} and a \code{PropScore} column giving the estimated propensity score for each combination of their levels. This should have been returned by \code{propscore_df}.}

\item{seed}{The random seed to be used, for reproducibility}

\item{use_attr}{logical. Should the attributes of \code{propscore_df} (set by \code{\link[=propscore_df]{propscore_df()}}) be used to specify \code{cov_cols}, \code{arm_col} and \code{intervention_level}?}

\item{replace}{logical. Should the comparison cases be sampled with replacement?}

\item{downsample}{logical. Should the treatment cases be downsampled when there aren't enough comparison cases? Only one of \code{replace} and \code{downsample} should be TRUE.}

\item{drop_int}{logical. Should cases in the intervention/treated group be dropped if there are no equivalent cases in the comparison cohort to sample from?}

\item{cov_cols}{(only specify if \code{use_attr == FALSE}) A character vector containing the column names of the covariates to be matched on. These should all be factor / categorical data.}

\item{arm_col}{(only specify if \code{use_attr == FALSE}) The name of the column indicating which rows are treated cases and which are comparison cases. These should be factor or character, with only two levels / options.}

\item{intervention_level}{(only specify if \code{use_attr == FALSE}) The value in the \code{arm_col} for the treatment cases}
}
\value{
\code{\link[=matched_sample]{matched_sample()}} returns a data frame that is an expanded form of \code{df}, with the same number of rows, in the same order, but some extra columns:
\itemize{
\item \code{PropScore}: the propensity score for that row, from \code{propscore_df}
\item \code{include}: how many copies of this row are included in the matched dataset. Zero means the row has been dropped. Values greater than one mean the row will be duplicated.
\item \code{seed}: the random seed that was used. This will be the same for all rows and is included in the output for reproducibility.
\code{\link[=matched_sample]{matched_sample()}} will also issue messages indicating:
\item How many comparison cases have been repeated (and how many times) if \code{replace == TRUE}
\item How many intervention cases have been downsampled, if \code{downsample == TRUE}
\item How many (if any) intervention cases have been dropped because there are no similar comparison cases to sample from, if \code{drop_int == TRUE}
\item How many (if any) intervention cases have been kept even though there are no similar comparison cases to sample from, if \code{drop_int == TRUE}
}
}
\description{
Generate a matched sample using the propensity score sampling approach
}
\examples{

## First use `propscore_df` to create a propensity score data frame
## This example uses the joint approach and specifies p(Z=1) such
## that the expected size of the comparison group is 300

eg_j_exp300 = propscore_df(
   df = eg_data,
   cov_cols = c("risk_ass", "category", "sus_age_bin"),
   arm_col = "Arm",
   intervention_level = "Intervention",
   cov_dist = "joint",
   pz1_fn = pz1_expectedN(300)
   )

## Now we can use `eg_j_exp300` to find a matched comparison group
## In this example we use the attribute information in `propscore_df`,
## we sample from the comparison group with replacement
## and do not drop any intervention cases for which there are no
## comparison cases to sample
match_j_exp300_replace = matched_sample(
   df = eg_data,
   propscore_df = eg_j_exp300,
   seed = 20,
   replace = TRUE,
   downsample = FALSE,
   drop_int = FALSE
   )

}
\seealso{
\code{\link[=expand_matched_df]{expand_matched_df()}}
}
