% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss_check_data.R
\name{pss_check_data}
\alias{pss_check_data}
\title{Check data}
\usage{
pss_check_data(df, cov_cols, arm_col, intervention_level)
}
\arguments{
\item{df}{A data frame containing treatment group and comparison cohort data}

\item{cov_cols}{A vector of strings, the column names of the covariates to be matched on. These should all be factor / categorical data.}

\item{arm_col}{The name of the column indicating which rows are treated cases and which are comparison cases. These should be factor or character, with only two levels / options.}

\item{intervention_level}{The value in the \code{arm_col} for the treatment cases}
}
\value{
\code{pss_check_data} returns a (possibly slightly tidied up) data frame or an error message. The data frame will have an \code{Arm} column with levels \code{Intervention} and \code{Comparison}.
}
\description{
\code{pss_check_data} checks that the data frame provided is suitable. Will return either a very slightly cleaned up version, or an error message.
The function does the following:
\itemize{
\item checks the \code{arm_col} column has two unique values, one of which is \code{intervention_level}
\item adds an \code{Arm} column, with two levels ("Intervention" and "Comparison")
\item checks the columns named in \code{cov_cols} are present and in a suitable form (ideally factor, but if character it will coerce to factor and give a warning)
\item Removes rows for which any of \code{cov_cols} or \code{arm_col} is missing.
}
}
\examples{
pss_check_data(
  df = eg_data,
  cov_cols = c("risk_ass", "category", "sus_age_bin"),
  arm_col = "Arm",
  intervention_level = "Intervention"
)
}
